in case of crash

+++ lib/similarity_search/sequential_scan.cpp
@@ -3,16 +3,16 @@
 using std::vector;
 
 
-float seq_scan::mse_l2(const float* const s1start, const float* const s2start, unsigned int len)
+double seq_scan::mse_l2(const double* const s1start, const double* const s2start, unsigned int len)
 {
-  float mse = 0;
+  double mse = 0;
   for (int i=0; i < len; ++i) {
     mse += ( *(s1start+i) - *(s2start+i) ) * ( *(s1start+i) - *(s2start+i) );
   }
   return mse;
 }
 
-vector<int> seq_scan::find_similar_subseq_indexes(const vector<float>& series, const vector<float>& query, float epsilon)
+vector<int> seq_scan::find_similar_subseq_indexes(const vector<double>& series, const vector<double>& query, double epsilon)
 {
   if (series.size() <= query.size()) return {0};
   if (epsilon < 0) return {};
