+++ lib/similarity_search/sequential_scan.h
@@ -6,9 +6,9 @@
 
 
 namespace seq_scan {
-float mse_l2(const float* const s1start, const float* const s2start, unsigned int len);
+double mse_l2(const double* const s1start, const double* const s2start, unsigned int len);
 
-std::vector<int> find_similar_subseq_indexes(const std::vector<float>& series, const std::vector<float>& query, float epsilon);
+std::vector<int> find_similar_subseq_indexes(const std::vector<double>& series, const std::vector<double>& query, double epsilon);
 };
 
 #endif
